save last post's number in localStorage aside the date
Now threads.coffee takes track of the number of the latest
post read and, aside de-highlighting threads, it also changes
the hash of the last reply link to the latest post read.
This currently doesn't work if the last reply is on a different
page as the latest post read, since we only change the hash, not
the page number.

Moreover, the thread de-highlighting is now done via a CSS class
'seen'.